Created new room system
Created a new room system that allows players to create a room with a random room code, and then have other players join the room via the correct room code.
Players in the same room can chat with each other. Rooms are automatically destroyed when players leave, and if players are already in a room, they can't join another.

1.4 h